+++ forui/lib/src/foundation/tappable.dart
@@ -45,6 +45,7 @@ class FTappable extends StatefulWidget {
   final Duration touchHoverExitDuration;
   final VoidCallback? onPress;
   final VoidCallback? onLongPress;
+  final Duration shortPressDelay;
   final ValueWidgetBuilder<FTappableState> builder;
   final Widget? child;
 
@@ -77,6 +78,7 @@ class FTappable extends StatefulWidget {
     this.touchHoverExitDuration = Duration.zero,
     this.onPress,
     this.onLongPress,
+    this.shortPressDelay = const Duration(milliseconds: 200),
     ValueWidgetBuilder<FTappableState>? builder,
     this.child,
     super.key,
@@ -111,6 +113,7 @@ class FTappable extends StatefulWidget {
       ..add(DiagnosticsProperty('touchHoverExitDuration', touchHoverExitDuration, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('onPress', onPress, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('onLongPress', onLongPress, level: DiagnosticLevel.debug))
+      ..add(DiagnosticsProperty('shortPressDelay', shortPressDelay, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('builder', builder, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('child', child, level: DiagnosticLevel.debug));
   }
@@ -148,6 +151,7 @@ class _FTappableState extends State<FTappable> {
           child: Listener(
             onPointerDown: (_) async {
               final count = ++_monotonic;
+              onPointerDown();
               await Future.delayed(widget.touchHoverEnterDuration);
               if (mounted && count == _monotonic && !_touchHovered) {
                 setState(() => _touchHovered = true);
@@ -155,6 +159,7 @@ class _FTappableState extends State<FTappable> {
             },
             onPointerUp: (_) async {
               final count = ++_monotonic;
+              onPointerUp();
               await Future.delayed(widget.touchHoverExitDuration);
               if (mounted && count == _monotonic && _touchHovered) {
                 setState(() => _touchHovered = false);
@@ -171,17 +176,17 @@ class _FTappableState extends State<FTappable> {
     }
 
     return Shortcuts(
-      shortcuts: const {
-        SingleActivator(LogicalKeyboardKey.enter): ActivateIntent(),
-      },
+      shortcuts: const {SingleActivator(LogicalKeyboardKey.enter): ActivateIntent()},
       child: Actions(
-        actions: {
-          ActivateIntent: CallbackAction<ActivateIntent>(onInvoke: (_) => widget.onPress!()),
-        },
+        actions: {ActivateIntent: CallbackAction<ActivateIntent>(onInvoke: (_) => widget.onPress!())},
         child: tappable,
       ),
     );
   }
+
+  void onPointerDown() {}
+
+  void onPointerUp() {}
 
   Widget get _child => GestureDetector(
         behavior: widget.behavior,
@@ -221,25 +226,21 @@ class _AnimatedTappableState extends _FTappableState with SingleTickerProviderSt
   void initState() {
     super.initState();
     _controller = AnimationController(vsync: this, duration: const Duration(milliseconds: 100));
-    _animation = Tween(begin: 1.0, end: 0.97).animate(_controller)
-      ..addStatusListener((status) {
-        if (status == AnimationStatus.completed) {
-          _controller.reverse();
-        }
-      });
+    _animation = Tween(begin: 1.0, end: 0.97).animate(_controller);
   }
+
+  @override
+  void onPointerDown() => _controller.forward();
+
+  @override
+  void onPointerUp() => _controller.reverse();
 
   @override
   Widget get _child => ScaleTransition(
         scale: _animation,
         child: GestureDetector(
           behavior: widget.behavior,
-          onTap: widget.onPress == null
-              ? null
-              : () {
-                  widget.onPress!();
-                  _controller.forward();
-                },
+          onTap: widget.onPress,
           onLongPress: widget.onLongPress,
           child: widget.builder(context, (focused: _focused, hovered: _hovered || _touchHovered), widget.child),
         ),