+++ forui/test/src/widgets/calendar/calendar_test.dart
@@ -11,7 +11,7 @@ void main() {
       testWidgets('navigates to previous page', (tester) async {
         await tester.pumpWidget(
           TestScaffold(
-            data: FThemes.zinc.light,
+            theme: FThemes.zinc.light,
             child: FCalendar(
               controller: FCalendarController.dates(selectable: (date) => date != DateTime.utc(2024, 7, 2)),
               start: DateTime(1900, 1, 8),
@@ -32,7 +32,7 @@ void main() {
       testWidgets('did not navigate to previous page', (tester) async {
         await tester.pumpWidget(
           TestScaffold(
-            data: FThemes.zinc.light,
+            theme: FThemes.zinc.light,
             child: FCalendar(
               controller: FCalendarController.dates(selectable: (date) => date != DateTime.utc(2024, 7, 2)),
               start: DateTime(2024, 7),
@@ -55,7 +55,7 @@ void main() {
       testWidgets('navigates to next page', (tester) async {
         await tester.pumpWidget(
           TestScaffold(
-            data: FThemes.zinc.light,
+            theme: FThemes.zinc.light,
             child: FCalendar(
               controller: FCalendarController.dates(selectable: (date) => date != DateTime.utc(2024, 7, 2)),
               start: DateTime(1900, 1, 8),
@@ -76,7 +76,7 @@ void main() {
       testWidgets('did not navigate to next page', (tester) async {
         await tester.pumpWidget(
           TestScaffold(
-            data: FThemes.zinc.light,
+            theme: FThemes.zinc.light,
             child: FCalendar(
               controller: FCalendarController.dates(selectable: (date) => date != DateTime.utc(2024, 7, 2)),
               start: DateTime(2024),
@@ -100,7 +100,7 @@ void main() {
 
       setUp(() {
         calendar = TestScaffold(
-          data: FThemes.zinc.light,
+          theme: FThemes.zinc.light,
           child: FCalendar(
             initialType: FCalendarPickerType.yearMonth,
             controller: FCalendarController.dates(),
@@ -176,7 +176,7 @@ void main() {
         testWidgets('select year where current month out of range', (tester) async {
           await tester.pumpWidget(
             TestScaffold(
-              data: FThemes.zinc.light,
+              theme: FThemes.zinc.light,
               child: FCalendar(
                 initialType: FCalendarPickerType.yearMonth,
                 controller: FCalendarController.dates(),