+++ forui/test/src/widgets/select_group/select_group_golden_test.dart
@@ -23,15 +23,15 @@ void main() {
                   label: const Text('Select Group'),
                   description: const Text('Select Group Description'),
                   controller: FMultiSelectGroupController(values: {1}),
-                  items: [
+                  items: const [
                     FSelectGroupItem.checkbox(
                       value: 1,
-                      label: const Text('Checkbox 1'),
+                      label: Text('Checkbox 1'),
                       semanticLabel: 'Checkbox 1',
                     ),
                     FSelectGroupItem.radio(
                       value: 2,
-                      label: const Text('Checkbox 2'),
+                      label: Text('Checkbox 2'),
                       semanticLabel: 'Checkbox 2',
                     ),
                   ],
@@ -42,14 +42,14 @@ void main() {
         ),
       );
 
-      await expectLater(find.byType(TestScaffold), matchesGoldenFile('blue-screen.png'));
+      await expectLater(find.byType(TestScaffold), isBlueScreen);
     });
 
     for (final (name, theme, background) in TestScaffold.themes) {
       testWidgets('$name with checkbox', (tester) async {
         await tester.pumpWidget(
           TestScaffold(
-            data: theme,
+            theme: theme,
             background: background,
             child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
@@ -60,20 +60,20 @@ void main() {
                     label: const Text('Select Group'),
                     description: const Text('Select Group Description'),
                     controller: FMultiSelectGroupController(values: {1}),
-                    items: [
+                    items: const [
                       FSelectGroupItem.checkbox(
                         value: 1,
-                        label: const Text('Checkbox 1'),
+                        label: Text('Checkbox 1'),
                         semanticLabel: 'Checkbox 1',
                       ),
                       FSelectGroupItem.checkbox(
                         value: 2,
-                        label: const Text('Checkbox 2'),
+                        label: Text('Checkbox 2'),
                         semanticLabel: 'Checkbox 2',
                       ),
                       FSelectGroupItem.checkbox(
                         value: 3,
-                        label: const Text('Checkbox 3'),
+                        label: Text('Checkbox 3'),
                         semanticLabel: 'Checkbox 3',
                       ),
                     ],
@@ -84,13 +84,13 @@ void main() {
           ),
         );
 
-        await expectLater(find.byType(TestScaffold), matchesGoldenFile('select-group/$name-checkbox.png'));
+        await expectLater(find.byType(TestScaffold), matchesGoldenFile('select-group/$name/checkbox.png'));
       });
 
       testWidgets('$name with checkbox error', (tester) async {
         await tester.pumpWidget(
           TestScaffold(
-            data: theme,
+            theme: theme,
             background: background,
             child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
@@ -102,20 +102,20 @@ void main() {
                     description: const Text('Select Group Description'),
                     forceErrorText: 'Some error message.',
                     controller: FMultiSelectGroupController(values: {1}),
-                    items: [
+                    items: const [
                       FSelectGroupItem.checkbox(
                         value: 1,
-                        label: const Text('Checkbox 1'),
+                        label: Text('Checkbox 1'),
                         semanticLabel: 'Checkbox 1',
                       ),
                       FSelectGroupItem.checkbox(
                         value: 2,
-                        label: const Text('Checkbox 2'),
+                        label: Text('Checkbox 2'),
                         semanticLabel: 'Checkbox 2',
                       ),
                       FSelectGroupItem.checkbox(
                         value: 3,
-                        label: const Text('Checkbox 3'),
+                        label: Text('Checkbox 3'),
                         semanticLabel: 'Checkbox 3',
                       ),
                     ],
@@ -126,7 +126,7 @@ void main() {
           ),
         );
 
-        await expectLater(find.byType(TestScaffold), matchesGoldenFile('select-group/$name-checkbox-error.png'));
+        await expectLater(find.byType(TestScaffold), matchesGoldenFile('select-group/$name/checkbox-error.png'));
       });
     }
   });
@@ -136,7 +136,7 @@ void main() {
       testWidgets('$name with radio', (tester) async {
         await tester.pumpWidget(
           TestScaffold(
-            data: theme,
+            theme: theme,
             background: background,
             child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
@@ -147,20 +147,20 @@ void main() {
                     label: const Text('Select Group'),
                     description: const Text('Select Group Description'),
                     controller: FRadioSelectGroupController(value: 1),
-                    items: [
+                    items: const [
                       FSelectGroupItem.radio(
                         value: 1,
-                        label: const Text('Radio 1'),
+                        label: Text('Radio 1'),
                         semanticLabel: 'Radio 1',
                       ),
                       FSelectGroupItem.radio(
                         value: 2,
-                        label: const Text('Radio 2'),
+                        label: Text('Radio 2'),
                         semanticLabel: 'Radio 2',
                       ),
                       FSelectGroupItem.radio(
                         value: 3,
-                        label: const Text('Radio 3'),
+                        label: Text('Radio 3'),
                         semanticLabel: 'Radio 3',
                       ),
                     ],
@@ -171,13 +171,13 @@ void main() {
           ),
         );
 
-        await expectLater(find.byType(TestScaffold), matchesGoldenFile('select-group/$name-radio.png'));
+        await expectLater(find.byType(TestScaffold), matchesGoldenFile('select-group/$name/radio.png'));
       });
 
       testWidgets('$name with radio error', (tester) async {
         await tester.pumpWidget(
           TestScaffold(
-            data: theme,
+            theme: theme,
             background: background,
             child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
@@ -189,20 +189,20 @@ void main() {
                     description: const Text('Select Group Description'),
                     forceErrorText: 'Some error message.',
                     controller: FRadioSelectGroupController(value: 1),
-                    items: [
+                    items: const [
                       FSelectGroupItem.radio(
                         value: 1,
-                        label: const Text('Radio 1'),
+                        label: Text('Radio 1'),
                         semanticLabel: 'Radio 1',
                       ),
                       FSelectGroupItem.radio(
                         value: 2,
-                        label: const Text('Radio 2'),
+                        label: Text('Radio 2'),
                         semanticLabel: 'Radio 2',
                       ),
                       FSelectGroupItem.radio(
                         value: 3,
-                        label: const Text('Radio 3'),
+                        label: Text('Radio 3'),
                         semanticLabel: 'Radio 3',
                       ),
                     ],
@@ -213,7 +213,7 @@ void main() {
           ),
         );
 
-        await expectLater(find.byType(TestScaffold), matchesGoldenFile('select-group/$name-radio-error.png'));
+        await expectLater(find.byType(TestScaffold), matchesGoldenFile('select-group/$name/radio-error.png'));
       });
     }
   });