+++ forui/CHANGELOG.md
@@ -1,8 +1,17 @@
-## 0.7.0
+## 0.7.0 (Next)
 
 ### Additions
 
 * Add `FButtonSpinner`.
+
+* Add `FTile`.
+
+* Add `FTileGroup`.
+
+### Fixes
+
+* Change FButton's animation to only start on mouse down and up.
+  [Reddit Comment](https://www.reddit.com/r/FlutterDev/comments/1g5tmu5/comment/lsm4gsd/?utm_source=share&utm_medium=web3x&utm_name=web3xcss&utm_term=1&utm_content=share_button).
 
 ## 0.6.1